cleanup

+++ cute_kernels/kernels/add/add_scalar/cuda_implementation/forward.cu
@@ -8,7 +8,7 @@
 #include "../../../../include/threads.h"
 
 template <typename scalar_t>
-__global__ void _add_scalar_cuda_kernel(const scalar_t *x, const fp32 y, scalar_t *output, const uint64 num_elements) {
+__global__ void _add_scalar_cuda_kernel(const scalar_t *x, const fp32 y, scalar_t *output, const uint32 num_elements) {
     constexpr int num_elements_per_thread = 16 / sizeof(scalar_t);
     static_assert(num_elements_per_thread == 4 || num_elements_per_thread == 8);
 
@@ -63,8 +63,8 @@ void add_scalar_cuda(const torch::Tensor &x, const float &y, torch::Tensor &outp
                 ChunkedArray<scalar_t> x_chunk = x_chunks[i];
                 ChunkedArray<scalar_t> output_chunk = output_chunks[i];
 
-                const uint64 num_elements = x_chunk.num_elements;
-                const uint32 NUM_BLOCKS = ceil_divide<uint64>(num_elements, num_elements_per_block);
+                const uint32 num_elements = x_chunk.num_elements;
+                const uint32 NUM_BLOCKS = ceil_divide<uint32>(num_elements, num_elements_per_block);
 
                 if constexpr (std::is_same_v<scalar_t, fp32>) {
                     _add_scalar_cuda_kernel<scalar_t>
